PHPStan level 9
If `DbService::readTextContent()` fails to read the file for whatever
reason, it returns `null`.  We could inform the user about that, but
rather assume that this happened because there is no moved.txt yet, and
let the user proceed editing the contents.  A possible failure when
saving will be reported, so users still can react.